# docs/api/tracker-openapi.yaml
openapi: 3.0.3
info:
  title: Martha Tracker API
  description: |
    Comprehensive REST API for the Martha multi-board tracker system.

    The tracker provides hierarchical issue management across multiple worktrees and boards,
    with support for epics, stories, tasks, and bugs. It includes rich metadata, documentation
    linking, agent integration, and real-time collaboration features.

    ## Hierarchical Structure

    The tracker follows a three-level hierarchy:
    - **Worktrees**: Top-level containers representing code repositories or projects
    - **Boards**: Kanban boards within a worktree for organizing issues
    - **Issues**: Individual work items (epics, stories, tasks, bugs)

    ## API Patterns

    - **Hierarchical Routes**: `/api/tracker/worktrees/{worktreeId}/boards/{boardId}/issues`
    - **Legacy Routes**: `/api/tracker/issues` (backward compatible)
    - **Error Format**: Consistent JSON error responses
    - **Authentication**: Cookie-based session auth (credentials: include)
  version: 1.0.0
  contact:
    name: Martha Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:20000/api/tracker
    description: Local development server
  - url: https://api.martha.dev/api/tracker
    description: Production server
...
components:
  schemas:
...
    Issue:
      type: object
      required:
        - id
        - worktree_id
        - board_id
        - type
        - title
        - description
        - status
        - priority
        - labels
        - documentation
        - links
        - metadata
      properties:
        id:
          type: string
          description: Unique issue identifier
          example: issue-001
        worktree_id:
          type: string
          description: Parent worktree ID
          example: martha-dev-v4
        board_id:
          type: string
          description: Parent board ID
          example: phase1-temporal-foundation
        type:
          type: string
          enum: [epic, story, task, bug]
          description: Issue type
          example: task
        title:
          type: string
          description: Issue title
          example: Implement temporal workflow engine
        description:
          type: string
          description: Detailed issue description (markdown)
          example: |
            Build the core temporal workflow execution engine with support for:
            - Activity scheduling
            - Timer management
            - Signal handling
        status:
          type: string
          description: Current status (matches board column ID)
          example: in-progress
        priority:
          type: string
          enum: [critical, high, medium, low]
          description: Issue priority
          example: high
        parent_id:
          type: string
          nullable: true
          description: Parent issue ID (for subtasks under epics/stories)
          example: epic-temporal-foundation
        assignee:
          $ref: '#/components/schemas/Assignee'
        labels:
          type: array
          items:
            type: string
          description: Issue labels/tags
          example: ["backend", "temporal", "high-priority"]
        documentation:
          $ref: '#/components/schemas/DocumentationLinks'
        links:
          $ref: '#/components/schemas/Links'
        metadata:
          $ref: '#/components/schemas/Metadata'
    Assignee:
      type: object
      nullable: true
      required:
        - id
        - name
        - avatar
      properties:
        id:
          type: string
          example: user-123
        name:
          type: string
          example: John Doe
        avatar:
          type: string
          format: uri
          example: https://avatar.example.com/user-123.png
    DocumentationLinks:
      type: object
      required:
        - overview
        - technical_spec
        - related_docs
      properties:
        overview:
          type: string
          nullable: true
          description: Overview documentation ID
          example: doc-temporal-overview
        technical_spec:
          type: string
          nullable: true
          description: Technical specification document ID
          example: doc-temporal-spec
        related_docs:
          type: array
          items:
            type: string
          description: Related documentation IDs
          example: ["doc-workflow-guide", "doc-api-reference"]
    Links:
      type: object
      required:
        - pr
        - related_issues
        - external
      properties:
        pr:
          type: string
          nullable: true
          description: Pull request URL
          example: https://github.com/user/repo/pull/123
        related_issues:
          type: array
          items:
            type: string
          description: Related issue IDs
          example: ["issue-002", "issue-003"]
        external:
          type: array
          items:
            type: string
          description: External links
          example: ["https://docs.temporal.io/"]
    Metadata:
      type: object
      required:
        - created_at
        - updated_at
        - version
      properties:
        created_at:
          type: string
          format: date-time
          example: "2026-01-15T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-16T14:30:00Z"
        version:
          type: integer
          description: Issue version (for optimistic locking)
          example: 3
...
    IssueListResponse:
      type: object
      required:
        - issues
        - count
      properties:
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
        count:
          type: integer
...
paths:
...
  /worktrees/{worktreeId}/boards/{boardId}/issues:
    parameters:
      - name: worktreeId
        in: path
        required: true
        schema:
          type: string
      - name: boardId
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Issues (Hierarchical)
      summary: List issues in board
      description: Retrieves all issues in a board with optional filtering
      operationId: listIssuesHierarchical
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: type
          in: query
          schema:
            type: string
            enum: [epic, story, task, bug]
          description: Filter by type
        - name: parent_id
          in: query
          schema:
            type: string
          description: Filter by parent issue ID
        - name: assignee
          in: query
          schema:
            type: string
          description: Filter by assignee ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'

    post:
      tags:
        - Issues (Hierarchical)
      summary: Create issue
      description: Creates a new issue in the board
      operationId: createIssueHierarchical
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                type:
                  type: string
                  enum: [epic, story, task, bug]
                  default: task
                status:
                  type: string
                  default: backlog
                parent_id:
                  type: string
                  nullable: true
                description:
                  type: string
                  default: ""
                labels:
                  type: array
                  items:
                    type: string
                  default: []
                priority:
                  type: string
                  enum: [critical, high, medium, low]
                  default: medium
            example:
              title: Implement workflow execution engine
              type: task
              status: backlog
              parent_id: null
              description: Build the core workflow execution logic
              labels: ["backend", "temporal"]
              priority: high
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '400':
          description: Invalid request
  /worktrees/{worktreeId}/boards/{boardId}/issues/{issueId}:
    parameters:
      - name: worktreeId
        in: path
        required: true
        schema:
          type: string
      - name: boardId
        in: path
        required: true
        schema:
          type: string
      - name: issueId
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Issues (Hierarchical)
      summary: Get issue
      description: Retrieves a specific issue
      operationId: getIssueHierarchical
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '404':
          description: Issue not found

    patch:
      tags:
        - Issues (Hierarchical)
      summary: Update issue
      description: Updates issue fields
      operationId: updateIssueHierarchical
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                priority:
                  type: string
                  enum: [critical, high, medium, low]
                labels:
                  type: array
                  items:
                    type: string
                assignee:
                  $ref: '#/components/schemas/Assignee'
            example:
              title: Implement workflow execution engine (updated)
              priority: critical
      responses:
        '200':
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '404':
          description: Issue not found

    delete:
      tags:
        - Issues (Hierarchical)
      summary: Delete issue
      description: Deletes an issue
      operationId: deleteIssueHierarchical
      responses:
        '204':
          description: Issue deleted
        '404':
          description: Issue not found
  /worktrees/{worktreeId}/boards/{boardId}/issues/{issueId}/move:
    parameters:
      - name: worktreeId
        in: path
        required: true
        schema:
          type: string
      - name: boardId
        in: path
        required: true
        schema:
          type: string
      - name: issueId
        in: path
        required: true
        schema:
          type: string

    post:
      tags:
        - Issues (Hierarchical)
      summary: Move issue
      description: Moves an issue to a different status/column
      operationId: moveIssueHierarchical
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  description: New status (column ID)
                index:
                  type: integer
                  description: Position in the new column (optional)
            example:
              status: in-progress
              index: 2
      responses:
        '200':
          description: Issue moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        '404':
          description: Issue not found

  # Legacy Issues API (Backward Compatible)
...
  /issues:
    get:
      tags:
        - Issues (Legacy)
      summary: List all issues
      description: Legacy endpoint - lists all issues across all boards
      operationId: listIssuesLegacy
      deprecated: true
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: parent_id
          in: query
          schema:
            type: string
        - name: assignee
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'

    post:
      tags:
        - Issues (Legacy)
      summary: Create issue
      description: Legacy endpoint - creates an issue in the default board
      operationId: createIssueLegacy
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                type:
                  type: string
                status:
                  type: string
                parent_id:
                  type: string
                description:
                  type: string
                labels:
                  type: array
                  items:
                    type: string
                priority:
                  type: string
      responses:
        '201':
          description: Issue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
  /issues/{issueId}:
    parameters:
      - name: issueId
        in: path
        required: true
        schema:
          type: string

    get:
      tags:
        - Issues (Legacy)
      summary: Get issue
      description: Legacy endpoint - retrieves an issue
      operationId: getIssueLegacy
      deprecated: true
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'

    patch:
      tags:
        - Issues (Legacy)
      summary: Update issue
      description: Legacy endpoint - updates an issue
      operationId: updateIssueLegacy
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Issue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'

    delete:
      tags:
        - Issues (Legacy)
      summary: Delete issue
      description: Legacy endpoint - deletes an issue
      operationId: deleteIssueLegacy
      deprecated: true
      responses:
        '204':
          description: Issue deleted
  /issues/{issueId}/move:
    parameters:
      - name: issueId
        in: path
        required: true
        schema:
          type: string

    post:
      tags:
        - Issues (Legacy)
      summary: Move issue
      description: Legacy endpoint - moves an issue
      operationId: moveIssueLegacy
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                index:
                  type: integer
      responses:
        '200':
          description: Issue moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'

  # Comments API
...